moved settings UI files one golder hierarchy level higher
