+++ openpype/tools/settings/settings/categories.py
