+++ openpype/tools/settings/settings/item_widgets.py
