+++ openpype/tools/settings/settings/list_item_widget.py
