+++ openpype/tools/settings/settings/list_strict_widget.py
