--- openpype/tools/settings/settings/widgets/__init__.py
@@ -1,8 +1,0 @@
-from .window import MainWidget
-from .widgets import ProjectListWidget
-
-
-__all__ = [
-    "MainWidget",
-    "ProjectListWidget"
-]
